create method in progress, have to stop here for today
current issue to solve - update all of the planets to have the solar system id which is being created. the view will list all existing planets and each will have tick box. each tick box contents are to be added into a list. this selected list of planets is then processed for individual editing. via a foreach, so that each planet has the id on them. it is probably better to do it the other way around, but meh.