added PlayerProfile and TitanOwnership models
playerprofile holds specific instances of titanownerships, which are the actual titans players will use. The titans themselves are now as a "template" properties of which are inherited into the titanownership model. at this point, i need to do a big overhaul of the 2 front ends of the game. for player profile testing purposes i need the game to have a galaxy, solar systems, planets, titans and users. This means that, right now, i cannot continue in this branch, i need to finish planets branch, and proceed to merge and integrate codebases of both branches. big times ahead.